backend-challenge-v1/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic[email]==2.12.4
pwdlib[argon2]==0.3.0
argon2-cffi==25.1.0
bcrypt>=4.0.0
PyJWT==2.10.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

